Extract a shared Metal graph-execution test harness
Graph Execution, Routing Nodes, Consolidated Numeric Nodes, Dictionary Ports and Graph Export Renderer each carried their own near-identical private harness for bringing up a device/Context/GraphRenderer, and the five copies had quietly drifted apart — different default render sizes (320x180 vs 64x64), inconsistent commandBuffer.error checking (most suites checked it, Consolidated Numeric Nodes silently swallowed GPU errors, Dictionary Ports never committed a command buffer at all), and five separate private test-failure error types.

These are now one GraphExecutionTestHarness and one GraphExecutionTestFailure in the new Tests/GraphExecutionTestSupport.swift, with render size and commandBuffer error checking exposed as explicit parameters so each suite's original behavior is preserved rather than quietly unified away.

Via Claude Sonnet 5
