+++ Tests/GraphExportRendererTests.swift
@@ -4,48 +4,6 @@ import Metal
 @testable import Fabric
 import Satin
 
-private struct GraphExportTestFailure: Error, CustomStringConvertible {
-    let description: String
-
-    init(_ description: String) {
-        self.description = description
-    }
-}
-
-private struct GraphExportTestHarness {
-    let context: Context
-
-    init?() {
-        guard let device = MTLCreateSystemDefaultDevice() else {
-            return nil
-        }
-
-        self.context = Context(
-            device: device,
-            sampleCount: 1,
-            colorPixelFormat: .bgra8Unorm,
-            depthPixelFormat: .depth32Float,
-            stencilPixelFormat: .invalid
-        )
-    }
-
-    func makeTexture(width: Int = 320, height: Int = 180, pixelFormat: MTLPixelFormat = .bgra8Unorm) throws -> MTLTexture {
-        let descriptor = MTLTextureDescriptor.texture2DDescriptor(
-            pixelFormat: pixelFormat,
-            width: width,
-            height: height,
-            mipmapped: false
-        )
-        descriptor.usage = [.renderTarget, .shaderRead]
-
-        guard let texture = self.context.device.makeTexture(descriptor: descriptor) else {
-            throw GraphExportTestFailure("Failed to create export texture")
-        }
-
-        return texture
-    }
-}
-
 private func publishExportPort(_ port: Fabric.Port, in graph: Graph) {
     port.published = true
     graph.rebuildPublishedParameterGroup()
@@ -53,7 +11,7 @@ private func publishExportPort(_ port: Fabric.Port, in graph: Graph) {
 
 private func requireExportValue<T>(_ value: T?, _ message: String) throws -> T {
     guard let value else {
-        throw GraphExportTestFailure(message)
+        throw GraphExecutionTestFailure(message)
     }
 
     return value
@@ -69,7 +27,7 @@ struct GraphExportRendererTests {
 
     @Test("Single frame export uses explicit graph time")
     func singleFrameExportUsesExplicitGraphTime() throws {
-        guard let harness = GraphExportTestHarness() else { return }
+        guard let harness = GraphExecutionTestHarness() else { return }
 
         let graph = Graph(context: harness.context)
         let timeNode = CurrentTimeNode(context: harness.context)
@@ -92,7 +50,7 @@ struct GraphExportRendererTests {
 
     @Test("Sequential export frames derive delta time internally")
     func sequentialExportFramesDeriveDeltaTimeInternally() throws {
-        guard let harness = GraphExportTestHarness() else { return }
+        guard let harness = GraphExecutionTestHarness() else { return }
 
         let graph = Graph(context: harness.context)
         let integralNode = NumberIntegralNode(context: harness.context)
@@ -122,7 +80,7 @@ struct GraphExportRendererTests {
 
     @Test("Sequential export frames advance internal frame number")
     func sequentialExportFramesAdvanceInternalFrameNumber() throws {
-        guard let harness = GraphExportTestHarness() else { return }
+        guard let harness = GraphExecutionTestHarness() else { return }
 
         let graph = Graph(context: harness.context)
         let renderInfoNode = RenderInfoNode(context: harness.context)
